Simplify the scheduling algorithm
The old scheduling code would try to find the best tile to place some
node on, which is complicated because it requires rolling back
scheduling decisions.  I never really made it far enough to get that
to work and have instead simply made the algorithm greedier (it has a
simple ordering for where an operation should be scheduler), which
removes the need to roll back decisions.

This gets rid of the lasts bits of cruft that were related to that old
scheduler.

+++ src/avail_mem.c++
@@ -30,7 +30,7 @@ avail_mem::avail_mem(size_t start_cycle, const std::weak_ptr<tile>& owner)
 std::string avail_mem::to_string(void) const
 {
     char buffer[1024];
-    if (_end_cycle == (size_t)(-1)) {
+    if (_end_cycle == -1) {
         snprintf(buffer, 1024, "memory in %s after %lu",
                  _owner.lock()->name().c_str(),
                  _start_cycle
